Fix shader precision where mediump (fp16 on mobile) loses data
- relief sampler: highp in both terrain stages; its texels are
  world-space heights, and fp16 rounding gave up to +/-2.3 m of
  per-texel vertical jitter above ~4200 m
- sky: compute the view ray and cos_gamma in highp; the sun-disc
  cutoff window (cos_gamma > 0.9995) is narrower than the fp16 ulp
  near 1.0, giving a ragged disc and banded halo. The v_uv varying
  must be highp in both stages: interpolation happens at the minimum
  of the two declared precisions
- ao_bake: drop the mediump annotations on the IGN dither, which
  collapsed to ~16 distinct values at fp16 and left directional
  banding in the baked ambient occlusion

+++ src/web/shaders/radial_common.vert
@@ -13,7 +13,7 @@ uniform highp vec2 inv_delta;
 uniform highp float inv_w;
 uniform highp float inv_avg_delta;
 uniform highp int max_lod;
-uniform mediump sampler2D relief;
+uniform highp sampler2D relief;
 
 // Output structure for radial vertex computation
 struct RadialVertex {
